[PROJ-85] refactor: 자산 관리 컴포넌트 변수명 일부 수정
